--- a/ggml/src/ggml-cuda/conv2d.cuh
+++ b/ggml/src/ggml-cuda/conv2d.cuh
@@ -5,6 +5,8 @@ constexpr int BS_OC     = 128;
 constexpr int BS_ICKHKW = 16;
 constexpr int BS_NOHOW  = 128;
 
+// supported configuration
+// https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#element-types-and-matrix-sizes
 constexpr int WMMA_M = 16;
 constexpr int WMMA_N = 16;
 constexpr int WMMA_K = 16;