--- a/ggml/src/ggml-cuda/conv2d.cuh
+++ b/ggml/src/ggml-cuda/conv2d.cuh
@@ -1,14 +1,12 @@
 #pragma once
 #include "common.cuh"
 
-constexpr int BS_OC     = 128;
-constexpr int BS_ICKHKW = 16;
-constexpr int BS_NOHOW  = 128;
+#define BS_OC     64
+#define BS_ICKHKW 16
+#define BS_NOHOW  64
 
-// supported configuration
-// https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#element-types-and-matrix-sizes
-constexpr int WMMA_M = 16;
-constexpr int WMMA_N = 16;
-constexpr int WMMA_K = 16;
+#define WMMA_M 16
+#define WMMA_N 16
+#define WMMA_K 16
 
 void ggml_cuda_op_conv2d(ggml_backend_cuda_context & ctx, ggml_tensor * dst);