@Green-Sky 2025-09-05
It is becoming increasingly hard to test these kind of changes with `sd.cpp`, @ggerganov please sync `ggml`, it has been over 3 weeks of rapid convolution development. :)