@mnehete32 2025-09-05
I think the kernel was not able to launch complete threads, as launcher launches warps per each WMMA_M, WMMA_N, i will work with launch fewer threads per block, also with the `mma.cuh` , it looks like I don't need shared memory to store results, I haven't checked completely yet. Also looking into it.