@Green-Sky 2025-09-13
### 768x1024 sd1 fp16 vae:
| method | time | memory |
|--------|--------|--------|
| CUDA imcol+mul | ~1.68s | 4992.19 MB |
| CUDA direct (master) | ~35.35s | 1920.19 MB |
| CUDA direct (ac5e0c0) (this pr) | ~9.00s | 1920.19 MB |
| CUDA implicitgemm (2ec76aa) | ~2.20s | 1920.19 MB |
| VULKAN imcol+mul | OOM | ~4992 MB |
| VULKAN direct | ~1.17s | 1920.19 MB |

Here is the table from #15805 extended with numbers from this pr.

I used the current pushed changes, you teased new numbers with better performance, but you still have not pushed those :) .