@JohannesGaessler 2025-09-14
>In my benchmarks FP16 actually slower than FP32. This is surprising since FP16 should normally be faster on Tensor Cores.
what I might be missing in my FP16 implementation?

The problem most likely has to do with either shared memory bank conflicts (you need a stride of 16 bytes between rows/columns) or excessive type conversions.