@JohannesGaessler 2025-09-14
Anyways, my criterion for approving and merging either one of the PRs as-is is for them to preserve existing functionality (as the tensor core PR would currently render some GPUs unusable) and to either be consistently faster than master or to add logic for selecting the optimal code paths depending on hardware and tensors.