@Green-Sky 2025-09-16
Updated table:

### 768x1024 sd1 fp16 vae:
| method | time | memory |
|--------|--------|--------|
| CUDA imcol+mul | ~1.68s | 4992.19 MB |
| CUDA direct (master) | ~35.35s | 1920.19 MB |
| CUDA direct (ac5e0c0) (this pr old) | ~9.00s | 1920.19 MB |
| CUDA direct (6049576) (this pr) | ~5.05s | 1920.19 MB |
| CUDA implicitgemm (2ec76aa) | ~2.20s | 1920.19 MB |
| VULKAN imcol+mul | OOM | ~4992 MB |
| VULKAN direct | ~1.17s | 1920.19 MB |