@etasnadi 2025-09-17
Hi, I just noticed this commit. 

Has anyone tested how the perf compares to the Vulkan implementation? There is also a coopmat2 impl that uses matrix cores.